requirements.txt
streamlit
tensorflow
pillow
opencv-python
numpy
